+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
+## 5.2.3
+- @uiw-admin/basic-layouts
+  - 🌟 `basic-layouts`：优化右上角menu关闭功能,通过useLayouts
 ## 5.2.1
 - @uiw-admin/basic-layouts
   - 🌟 `basic-layouts`：增加右上角菜单关闭功能